Restructure views and extract product data to a model
Moved view files under a consolidated "pages" directory for better organization. Adjusted imports accordingly and extracted hardcoded product data into a new `products` model file. This improves maintainability and promotes cleaner code structure.
