+++ src/view/components/loading-overlay/styles.ts
