+++ src/view/pages/menu-list/index.tsx
@@ -1,31 +1,8 @@
-import React, { useState } from 'react';
-import { View, FlatList, Image } from 'react-native';
-import { Text, Card, TextInput } from 'react-native-paper';
-import { styles } from './styles';
-
-const products = [
-  {
-    id: '1',
-    name: 'Pizza Margherita',
-    description: 'Molho de tomate, mussarela e manjericão fresco.',
-    price: 'R$ 35,00',
-    image: 'https://laticiniosbomdestino.com.br/2016/wp-content/uploads/2023/03/pizza-marguerita-com-mozzarella-de-bufala-bom-destino-scaled.jpg',
-  },
-  {
-    id: '2',
-    name: 'Hambúrguer Artesanal',
-    description: 'Pão brioche, carne 180g, queijo cheddar e bacon.',
-    price: 'R$ 28,00',
-    image: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQjF_0B9XbtUUI7pOZmO7TQyLtgcuQ4eujwXA&s',
-  },
-  {
-    id: '3',
-    name: 'Salada Caesar',
-    description: 'Alface, frango grelhado, parmesão e molho caesar.',
-    price: 'R$ 22,00',
-    image: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQHcqtOQGHJCk9x6BR8d0Coc7Xt1R9d0t-AlQ&s',
-  },
-];
+import React, {useState} from 'react';
+import {FlatList, Image, View} from 'react-native';
+import {Card, Text, TextInput} from 'react-native-paper';
+import {styles} from './styles';
+import {products} from "../../../model/products";
 
 export default function MenuList() {
   const [busca, setBusca] = useState('');
