+++ src/view/pages/onboarding/styles.ts
