+++ src/view/pages/product-registration/index.tsx
