+++ src/view/pages/product-registration/styles.ts
