+++ src/view/pages/restaurant-registration/index.tsx
