+++ src/view/pages/restaurant-registration/styles.ts
