+++ src/view/pages/user-registration/index.tsx
