+++ src/view/pages/user-registration/styles.ts
